Fix to Lazy Load Images ...
This will only load the Elected Official Images for when they are actually loaded into the page.  So mobile paying for data won't be charged for downloading dozens of 1024px images they may not even look at.  Should also make the page load faster, as the number of initial HTTP requests will drop.